--- catchBullying/Doctor/DoctorHomeViewController.swift
@@ -1,20 +1,0 @@
-//
-//  DoctorHomeViewController.swift
-//  catchBullying
-//
-//  Created by apple on 21/05/1443 AH.
-//
-
-import UIKit
-
-class DoctorHomeViewController: UITabBarController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-      self.navigationItem.setHidesBackButton(true, animated: false)
-    }
-    
-
-    
-}
