+++ catchBullying/ViewControllers/Calendar/DateSelectorViewController.swift
@@ -34,19 +34,6 @@ class DateSelectorViewController: UIViewController {
       fatalError()
     }
     
-    
-    
-    
   }
-  
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
 
 }
