+++ catchBullying/ViewControllers/Patient/UserHomeViewController.swift
@@ -1,0 +1,19 @@
+//
+//  UserHomeViewController.swift
+//  catchBullying
+//
+//  Created by apple on 21/05/1443 AH.
+//
+
+import UIKit
+
+class UserHomeViewController: UITabBarController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+      self.navigationItem.setHidesBackButton(true, animated: false)
+    }
+
+
+}
